requirements.txt
asgiref==3.6.0
autopep8==2.0.1
Django==3.2.16
django-cors-headers==3.13.0
djangorestframework==3.14.0
gunicorn==20.1.0
Pillow==9.4.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pytz==2022.7
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
whitenoise==6.3.0
